database seeder and factories created

+++ database/migrations/2022_08_04_071058_create_teams_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->foreignId('player_id');
             $table->foreignId('fixture_id');
             $table->string('league');
-            $table->int('points') ->default(0);
+            $table->tinyInteger('points')->default(0);
             $table->string('manager');
             $table->timestamps();
         });
